:recycle: feat: [FE] text atom 리팩토링
text atom 전체 프로젝트 구조에 맞게 리팩토링
