--- frontend/components/atoms/text/textInterface.ts
@@ -1,8 +1,0 @@
-export interface textProps {
-    text: string;
-    color?: string;
-}
-
-export interface textStyleProps {
-    color?: string;
-}
